requirements.txt
jupyter
ipykernel
notebook
colorama
ipdb
matplotlib
seaborn
numpy
pandas
scipy
tensorflow
opencv-python
# To balance classes with SMOTE
scikit-learn
# API
fastapi
uvicorn
pydantic
python-multipart
# dotenv
python-dotenv
